Añadida nueva sección de juegos (no está terminada)
La nueva ficha vendrá en el próximo commit pero para poder empezar a popular la
sección necesitamos estos cambios subidos.

Este commit también incluye las siguientes modificaciones:
- crear un juego ya no crea automáticamente una facción

- la habilidad de Crear Tag es ahora Crear Entidad y permite crear tags y crear
  juegos. Cualquiera que tenga la habilidad puede ir a Juegos > Crear juego e
  iniciar el proceso de creación de un juego (igual que tags).

- renombrado attributo code a slug (interno)

- Comprar facción ahora funciona solo para juegos (temporalmente) y requiere que
  el juego se haya creado de antemano

- Reducido a 3 número de contenidos iniciales para nuevos tags

- Añadidos nuevos atributos a juego:
  - fecha de lanzamiento
  - publisher
  - plataforma

+++ app/controllers/games_controller.rb
@@ -1,21 +1,48 @@
 # -*- encoding : utf-8 -*-
-class Admin::JuegosController < AdministrationController
+class GamesController < ApplicationController
+
+  def submenu
+    "Juegos"
+  end
+
+  def submenu_items
+    platforms = GamingPlatform.find(:all).sort_by{|platform|
+      platform.name.downcase
+    }.collect{|platform| [platform.name, "/juegos/p/#{platform.slug}"]}
+  end
 
   def index
+    platform = GamingPlatform.find_by_slug(params[:platform_code] || 'pc')
+
     @games = Game.paginate(
+        :conditions => ["gaming_platform_id = ?", platform.id],
         :order => 'LOWER(name)',
         :page => params[:page],
         :per_page => 50)
+    @title = "Juegos de #{platform.name}"
   end
 
   def new
+    require_authorization(:can_create_entities?)
     @game = Game.new
   end
 
   def create
+    require_authorization(:can_create_entities?)
+    params[:game][:user_id] = @user.id
     @game = Game.new(params[:game])
-    if @game.save
-      flash[:notice] = 'Juego creado correctamente.'
+    if @game.valid?
+      decision = Decision.create({
+        :decision_type_class => "CreateGame",
+        :context => {
+          :initiating_user_id => @user.id,
+          :game => params[:game],
+        },
+      })
+
+      flash[:notice] = (
+          "Solicitud para crear el juego <strong>#{@game.name}</strong> creado
+          correctamente.")
       redirect_to :action => 'index'
     else
       render :action => 'new'
@@ -24,6 +51,7 @@ class Admin::JuegosController < AdministrationController
 
   # TODO remove this from here
   def create_games_version
+    require_authorization(:can_create_entities?)
     gm = GamesVersion.new(params[:games_version])
     if gm.save
       flash[:notice] = 'Versión de juego creada correctamente.'
@@ -32,10 +60,11 @@ class Admin::JuegosController < AdministrationController
                       "#{gm.errors.full_messages_html}"
     end
 
-    redirect_to "/admin/juegos/edit/#{gm.game_id}"
+    redirect_to "/juegos/#{gm.game_id}/edit"
   end
 
   def create_games_mode
+    require_authorization(:can_create_entities?)
     gm = GamesMode.new(params[:games_mode])
     if gm.save
       flash[:notice] = 'Modo de juego creado correctamente.'
@@ -44,10 +73,11 @@ class Admin::JuegosController < AdministrationController
                       "#{gm.errors.full_messages_html}"
     end
 
-    redirect_to "/admin/juegos/edit/#{gm.game_id}"
+    redirect_to "/juegos/#{gm.game_id}/edit"
   end
 
   def destroy_games_version
+    require_authorization(:can_create_entities?)
     gv = GamesVersion.find(params[:id])
     if gv
       gv.destroy
@@ -56,10 +86,11 @@ class Admin::JuegosController < AdministrationController
       flash[:error] = "Error al borrar la versión: "+
                       "#{gv.errors.full_messages_html}"
     end
-    redirect_to "/admin/juegos/edit/#{gv.game_id}"
+    redirect_to "/juegos/#{gv.game_id}/edit"
   end
 
   def destroy_games_mode
+    require_authorization(:can_create_entities?)
     gv = GamesMode.find(params[:id])
     if gv
       gv.destroy
@@ -68,14 +99,20 @@ class Admin::JuegosController < AdministrationController
       flash[:error] = "Error al borrar el modo de juego: "+
                       "#{gv.errors.full_messages_html}"
     end
-    redirect_to "/admin/juegos/edit/#{gv.game_id}"
+    redirect_to "/juegos/#{gv.game_id}/edit"
   end
 
   def edit
+    require_authorization(:can_edit_entities?)
+    @game = Game.find(params[:id])
+  end
+
+  def show
     @game = Game.find(params[:id])
   end
 
   def update
+    require_authorization(:can_edit_entities?)
     @game = Game.find(params[:id])
     if @game.update_attributes(params[:game])
       flash[:notice] = 'Juego actualizado correctamente.'
@@ -85,10 +122,5 @@ class Admin::JuegosController < AdministrationController
                       "#{@game.errors.full_messages_html}"
       render :action => 'edit'
     end
-  end
-
-  def destroy
-    Game.find(params[:id]).destroy
-    redirect_to :action => 'index'
   end
 end
